+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/peervalidation/RequiredBlocksPeerValidator.java
@@ -16,35 +16,19 @@
  */
 package org.hyperledger.besu.ethereum.eth.peervalidation;
 
-import static com.google.common.base.Preconditions.checkArgument;
-
 import org.hyperledger.besu.ethereum.core.BlockHeader;
 import org.hyperledger.besu.ethereum.core.Hash;
-import org.hyperledger.besu.ethereum.eth.manager.EthContext;
 import org.hyperledger.besu.ethereum.eth.manager.EthPeer;
-import org.hyperledger.besu.ethereum.eth.manager.task.AbstractPeerTask;
-import org.hyperledger.besu.ethereum.eth.manager.task.GetHeadersFromPeerByNumberTask;
 import org.hyperledger.besu.ethereum.mainnet.ProtocolSchedule;
 import org.hyperledger.besu.plugin.services.MetricsSystem;
-
-import java.time.Duration;
-import java.util.List;
-import java.util.concurrent.CompletableFuture;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-public class RequiredBlocksPeerValidator implements PeerValidator {
+public class RequiredBlocksPeerValidator extends AbstractPeerBlockValidator {
   private static final Logger LOG = LogManager.getLogger();
-  private static long DEFAULT_CHAIN_HEIGHT_ESTIMATION_BUFFER = 10L;
 
-  private final ProtocolSchedule<?> protocolSchedule;
-  private final MetricsSystem metricsSystem;
-
-  private final long blockNumber;
   private final Hash hash;
-  // Wait for peer's chainhead to advance some distance beyond blockNumber before validating
-  private final long chainHeightEstimationBuffer;
 
   RequiredBlocksPeerValidator(
       final ProtocolSchedule<?> protocolSchedule,
@@ -52,12 +36,8 @@ public class RequiredBlocksPeerValidator implements PeerValidator {
       final long blockNumber,
       final Hash hash,
       final long chainHeightEstimationBuffer) {
-    checkArgument(chainHeightEstimationBuffer >= 0);
-    this.protocolSchedule = protocolSchedule;
-    this.metricsSystem = metricsSystem;
-    this.blockNumber = blockNumber;
+    super(protocolSchedule, metricsSystem, blockNumber, chainHeightEstimationBuffer);
     this.hash = hash;
-    this.chainHeightEstimationBuffer = chainHeightEstimationBuffer;
   }
 
   public RequiredBlocksPeerValidator(
@@ -70,63 +50,15 @@ public class RequiredBlocksPeerValidator implements PeerValidator {
   }
 
   @Override
-  public CompletableFuture<Boolean> validatePeer(
-      final EthContext ethContext, final EthPeer ethPeer) {
-    final AbstractPeerTask<List<BlockHeader>> getHeaderTask =
-        GetHeadersFromPeerByNumberTask.forSingleNumber(
-                protocolSchedule, ethContext, blockNumber, metricsSystem)
-            .setTimeout(Duration.ofSeconds(20))
-            .assignPeer(ethPeer);
-    return getHeaderTask
-        .run()
-        .handle(
-            (res, err) -> {
-              if (err != null) {
-                // Mark peer as invalid on error
-                LOG.debug(
-                    "Peer {} is invalid because required block block ({}) is unavailable: {}",
-                    ethPeer,
-                    blockNumber,
-                    err.toString());
-                return false;
-              }
-              final List<BlockHeader> headers = res.getResult();
-              if (headers.size() == 0) {
-                // If no headers are returned, fail
-                LOG.debug(
-                    "Peer {} is invalid because required block ({}) is unavailable.",
-                    ethPeer,
-                    blockNumber);
-                return false;
-              }
-              final BlockHeader header = headers.get(0);
-              final boolean validBlock = hash.equals(header.getHash());
-              if (!validBlock) {
-                LOG.debug(
-                    "Peer {} is invalid because required block ({}) does not match required hash ({}).",
-                    ethPeer,
-                    blockNumber,
-                    hash);
-              }
-              return validBlock;
-            });
-  }
-
-  @Override
-  public boolean canBeValidated(final EthPeer ethPeer) {
-    return ethPeer.chainState().getEstimatedHeight() >= (blockNumber + chainHeightEstimationBuffer);
-  }
-
-  @Override
-  public Duration nextValidationCheckTimeout(final EthPeer ethPeer) {
-    if (!ethPeer.chainState().hasEstimatedHeight()) {
-      return Duration.ofSeconds(30);
+  boolean validateBlockHeader(final EthPeer ethPeer, final BlockHeader header) {
+    final boolean validBlock = hash.equals(header.getHash());
+    if (!validBlock) {
+      LOG.debug(
+          "Peer {} is invalid because required block ({}) does not match required hash ({}).",
+          ethPeer,
+          blockNumber,
+          hash);
     }
-    final long distanceToBlock = blockNumber - ethPeer.chainState().getEstimatedHeight();
-    if (distanceToBlock < 100_000L) {
-      return Duration.ofMinutes(1);
-    }
-    // If the peer is trailing behind, give it some time to catch up before trying again.
-    return Duration.ofMinutes(10);
+    return validBlock;
   }
 }
