+++ ethereum/eth/src/test/java/org/hyperledger/besu/ethereum/eth/peervalidation/DaoForkPeerValidatorTest.java
@@ -19,54 +19,52 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.hyperledger.besu.ethereum.core.Block;
 import org.hyperledger.besu.ethereum.core.BlockDataGenerator;
 import org.hyperledger.besu.ethereum.core.BlockDataGenerator.BlockOptions;
-import org.hyperledger.besu.ethereum.eth.manager.DeterministicEthScheduler;
-import org.hyperledger.besu.ethereum.eth.manager.EthPeer;
 import org.hyperledger.besu.ethereum.eth.manager.EthProtocolManager;
 import org.hyperledger.besu.ethereum.eth.manager.EthProtocolManagerTestUtil;
 import org.hyperledger.besu.ethereum.eth.manager.RespondingEthPeer;
-import org.hyperledger.besu.ethereum.eth.messages.BlockHeadersMessage;
-import org.hyperledger.besu.ethereum.eth.messages.EthPV62;
-import org.hyperledger.besu.ethereum.eth.messages.GetBlockHeadersMessage;
 import org.hyperledger.besu.ethereum.mainnet.MainnetBlockHeaderValidator;
 import org.hyperledger.besu.ethereum.mainnet.MainnetProtocolSchedule;
 import org.hyperledger.besu.metrics.noop.NoOpMetricsSystem;
 import org.hyperledger.besu.util.bytes.BytesValue;
 
-import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import org.junit.Test;
 
-public class DaoForkPeerValidatorTest {
+public class DaoForkPeerValidatorTest extends AbstractPeerBlockValidatorTest {
+
+  @Override
+  AbstractPeerBlockValidator getSampleValidator(final long blockNumber) {
+    return new DaoForkPeerValidator(
+        MainnetProtocolSchedule.create(), new NoOpMetricsSystem(), blockNumber, 0);
+  }
 
   @Test
   public void validatePeer_responsivePeerOnRightSideOfFork() {
-    EthProtocolManager ethProtocolManager = EthProtocolManagerTestUtil.create();
-    BlockDataGenerator gen = new BlockDataGenerator(1);
-    long daoBlockNumber = 500;
-    Block daoBlock =
+    final EthProtocolManager ethProtocolManager = EthProtocolManagerTestUtil.create();
+    final BlockDataGenerator gen = new BlockDataGenerator(1);
+    final long daoBlockNumber = 500;
+    final Block daoBlock =
         gen.block(
             BlockOptions.create()
                 .setBlockNumber(daoBlockNumber)
                 .setExtraData(MainnetBlockHeaderValidator.DAO_EXTRA_DATA));
 
-    PeerValidator validator =
+    final PeerValidator validator =
         new DaoForkPeerValidator(
             MainnetProtocolSchedule.create(), new NoOpMetricsSystem(), daoBlockNumber, 0);
 
-    RespondingEthPeer peer =
+    final RespondingEthPeer peer =
         EthProtocolManagerTestUtil.createPeer(ethProtocolManager, daoBlockNumber);
 
-    CompletableFuture<Boolean> result =
+    final CompletableFuture<Boolean> result =
         validator.validatePeer(ethProtocolManager.ethContext(), peer.getEthPeer());
 
     assertThat(result).isNotDone();
 
     // Send response for dao block
-    AtomicBoolean daoBlockRequested = respondToDaoBlockRequest(peer, daoBlock);
+    final AtomicBoolean daoBlockRequested = respondToBlockRequest(peer, daoBlock);
 
     assertThat(daoBlockRequested).isTrue();
     assertThat(result).isDone();
@@ -75,149 +73,30 @@ public class DaoForkPeerValidatorTest {
 
   @Test
   public void validatePeer_responsivePeerOnWrongSideOfFork() {
-    EthProtocolManager ethProtocolManager = EthProtocolManagerTestUtil.create();
-    BlockDataGenerator gen = new BlockDataGenerator(1);
-    long daoBlockNumber = 500;
-    Block daoBlock =
+    final EthProtocolManager ethProtocolManager = EthProtocolManagerTestUtil.create();
+    final BlockDataGenerator gen = new BlockDataGenerator(1);
+    final long daoBlockNumber = 500;
+    final Block daoBlock =
         gen.block(
             BlockOptions.create().setBlockNumber(daoBlockNumber).setExtraData(BytesValue.EMPTY));
 
-    PeerValidator validator =
+    final PeerValidator validator =
         new DaoForkPeerValidator(
             MainnetProtocolSchedule.create(), new NoOpMetricsSystem(), daoBlockNumber, 0);
 
-    RespondingEthPeer peer =
+    final RespondingEthPeer peer =
         EthProtocolManagerTestUtil.createPeer(ethProtocolManager, daoBlockNumber);
 
-    CompletableFuture<Boolean> result =
+    final CompletableFuture<Boolean> result =
         validator.validatePeer(ethProtocolManager.ethContext(), peer.getEthPeer());
 
     assertThat(result).isNotDone();
 
     // Send response for dao block
-    AtomicBoolean daoBlockRequested = respondToDaoBlockRequest(peer, daoBlock);
+    final AtomicBoolean daoBlockRequested = respondToBlockRequest(peer, daoBlock);
 
     assertThat(daoBlockRequested).isTrue();
     assertThat(result).isDone();
     assertThat(result).isCompletedWithValue(false);
-  }
-
-  @Test
-  public void validatePeer_unresponsivePeer() {
-    EthProtocolManager ethProtocolManager =
-        EthProtocolManagerTestUtil.create(DeterministicEthScheduler.TimeoutPolicy.ALWAYS_TIMEOUT);
-    long daoBlockNumber = 500;
-
-    PeerValidator validator =
-        new DaoForkPeerValidator(
-            MainnetProtocolSchedule.create(), new NoOpMetricsSystem(), daoBlockNumber, 0);
-
-    RespondingEthPeer peer =
-        EthProtocolManagerTestUtil.createPeer(ethProtocolManager, daoBlockNumber);
-
-    CompletableFuture<Boolean> result =
-        validator.validatePeer(ethProtocolManager.ethContext(), peer.getEthPeer());
-
-    // Request should timeout immediately
-    assertThat(result).isDone();
-    assertThat(result).isCompletedWithValue(false);
-  }
-
-  @Test
-  public void validatePeer_requestBlockFromPeerBeingTested() {
-    EthProtocolManager ethProtocolManager = EthProtocolManagerTestUtil.create();
-    BlockDataGenerator gen = new BlockDataGenerator(1);
-    long daoBlockNumber = 500;
-    Block daoBlock =
-        gen.block(
-            BlockOptions.create()
-                .setBlockNumber(daoBlockNumber)
-                .setExtraData(MainnetBlockHeaderValidator.DAO_EXTRA_DATA));
-
-    PeerValidator validator =
-        new DaoForkPeerValidator(
-            MainnetProtocolSchedule.create(), new NoOpMetricsSystem(), daoBlockNumber, 0);
-
-    int peerCount = 1000;
-    List<RespondingEthPeer> otherPeers =
-        Stream.generate(
-                () -> EthProtocolManagerTestUtil.createPeer(ethProtocolManager, daoBlockNumber))
-            .limit(peerCount)
-            .collect(Collectors.toList());
-    RespondingEthPeer targetPeer =
-        EthProtocolManagerTestUtil.createPeer(ethProtocolManager, daoBlockNumber);
-
-    CompletableFuture<Boolean> result =
-        validator.validatePeer(ethProtocolManager.ethContext(), targetPeer.getEthPeer());
-
-    assertThat(result).isNotDone();
-
-    // Other peers should not receive request for dao block
-    for (RespondingEthPeer otherPeer : otherPeers) {
-      AtomicBoolean daoBlockRequestedForOtherPeer = respondToDaoBlockRequest(otherPeer, daoBlock);
-      assertThat(daoBlockRequestedForOtherPeer).isFalse();
-    }
-
-    // Target peer should receive request for dao block
-    final AtomicBoolean daoBlockRequested = respondToDaoBlockRequest(targetPeer, daoBlock);
-    assertThat(daoBlockRequested).isTrue();
-  }
-
-  @Test
-  public void canBeValidated() {
-    BlockDataGenerator gen = new BlockDataGenerator(1);
-    EthProtocolManager ethProtocolManager =
-        EthProtocolManagerTestUtil.create(DeterministicEthScheduler.TimeoutPolicy.ALWAYS_TIMEOUT);
-    long daoBlockNumber = 500;
-    long buffer = 10;
-
-    PeerValidator validator =
-        new DaoForkPeerValidator(
-            MainnetProtocolSchedule.create(), new NoOpMetricsSystem(), daoBlockNumber, buffer);
-
-    EthPeer peer = EthProtocolManagerTestUtil.createPeer(ethProtocolManager, 0).getEthPeer();
-
-    peer.chainState().update(gen.hash(), daoBlockNumber - 10);
-    assertThat(validator.canBeValidated(peer)).isFalse();
-
-    peer.chainState().update(gen.hash(), daoBlockNumber);
-    assertThat(validator.canBeValidated(peer)).isFalse();
-
-    peer.chainState().update(gen.hash(), daoBlockNumber + buffer - 1);
-    assertThat(validator.canBeValidated(peer)).isFalse();
-
-    peer.chainState().update(gen.hash(), daoBlockNumber + buffer);
-    assertThat(validator.canBeValidated(peer)).isTrue();
-
-    peer.chainState().update(gen.hash(), daoBlockNumber + buffer + 10);
-    assertThat(validator.canBeValidated(peer)).isTrue();
-  }
-
-  private AtomicBoolean respondToDaoBlockRequest(
-      final RespondingEthPeer peer, final Block daoBlock) {
-    AtomicBoolean daoBlockRequested = new AtomicBoolean(false);
-
-    RespondingEthPeer.Responder responder =
-        RespondingEthPeer.targetedResponder(
-            (cap, msg) -> {
-              if (msg.getCode() != EthPV62.GET_BLOCK_HEADERS) {
-                return false;
-              }
-              GetBlockHeadersMessage headersRequest = GetBlockHeadersMessage.readFrom(msg);
-              boolean isDaoBlockRequest =
-                  headersRequest.blockNumber().isPresent()
-                      && headersRequest.blockNumber().getAsLong()
-                          == daoBlock.getHeader().getNumber();
-              if (isDaoBlockRequest) {
-                daoBlockRequested.set(true);
-              }
-              return isDaoBlockRequest;
-            },
-            (cap, msg) -> BlockHeadersMessage.create(daoBlock.getHeader()));
-
-    // Respond
-    peer.respond(responder);
-
-    return daoBlockRequested;
   }
 }
